Extract formatting feature
