+++ __init__.py
@@ -3,3 +3,4 @@ from .core.panels import *
 from .features.completion import *
 from .features.diagnostics import *
 from .features.configuration import *
+from .features.formatting import *
